feat: migrate stylelint config to esm and fix related issues
