+++ app/package.json
@@ -71,6 +71,8 @@
         "@types/react-dom": "^18.0.11",
         "@types/sanitize-html": "^2.9.0",
         "@types/yargs": "^17.0.32",
+        "@typescript-eslint/eslint-plugin": "^8.11.0",
+        "@typescript-eslint/parser": "^8.11.0",
         "@vitejs/plugin-react-swc": "^3.5.0",
         "@vitest/coverage-v8": "^1.2.2",
         "autoprefixer": "^10.4.14",
