+++ app/src/components/Navbar/styles.module.css
@@ -4,8 +4,8 @@
     padding: var(--go-ui-spacing-md);
     max-width: 32rem;
     min-height: 16rem;
-    gap: var(--go-ui-spacing-md);
     overflow: auto;
+    gap: var(--go-ui-spacing-md);
 
     .option-list {
         flex-basis: 10rem;
@@ -75,7 +75,7 @@
 
             .brand {
                 display: flex;
-                align-items: top;
+                align-items: flex-start;
                 flex-grow: 1;
                 flex-wrap: wrap;
                 gap: var(--go-ui-spacing-sm);
