+++ app/src/views/CountryNsOverviewContextAndStructure/NationalSocietyLocalUnits/LocalUnitsMap/styles.module.css
@@ -28,7 +28,7 @@
             }
 
             .email {
-                word-break: break-word;
+                overflow-wrap: break-word;
             }
         }
     }
