+++ app/src/views/CountryOngoingActivitiesThreeWActivities/ResponseActivitiesMap/styles.module.css
@@ -45,9 +45,3 @@
         }
     }
 }
-
-.map-popup-content {
-    display: flex;
-    flex-direction: column;
-    gap: var(--go-ui-spacing-md);
-}
