+++ app/src/views/CountryOngoingActivitiesThreeWActivities/styles.module.css
@@ -68,10 +68,3 @@
         font-size: var(--go-ui-height-icon-multiplier);
     }
 }
-
-.tooltip {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    padding: var(--go-ui-spacing-sm);
-}
