+++ app/src/views/CountryPreparedness/DocumentCard/styles.module.css
@@ -14,7 +14,7 @@
 
         .name {
             display: flex;
-            word-break: break-word;
+            overflow-wrap: break-word;
         }
     }
