+++ app/src/views/DrefApplicationForm/EventDetail/styles.module.css
@@ -1,5 +1,5 @@
 .event-detail {
     display: flex;
-    flex-direction: column;
     gap: var(--go-ui-spacing-2xl);
+    flex-direction: column;
 }
