+++ app/src/views/EmergencyActivities/ActivityDetail/styles.module.css
@@ -48,7 +48,3 @@
         }
     }
 }
-
-.tooltip {
-    padding: var(--go-ui-spacing-sm);
-}
