+++ app/src/views/EmergencyActivities/styles.module.css
@@ -69,10 +69,3 @@
         font-size: var(--go-ui-height-icon-multiplier);
     }
 }
-
-.tooltip {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    padding: var(--go-ui-spacing-sm);
-}
