--- app/stylelint.config.cjs
@@ -1,38 +1,0 @@
-const path = require('path');
-
-const cssPaths = [
-    path.resolve(__dirname, './src/index.css'),
-    path.resolve(__dirname, './node_modules/@ifrc-go/ui/dist/index.css'),
-];
-
-module.exports = {
-    plugins: [
-        'stylelint-no-unused-selectors',
-        'stylelint-value-no-unknown-custom-properties',
-    ],
-    extends: [
-        'stylelint-config-recommended',
-        'stylelint-config-concentric',
-    ],
-    rules: {
-        // https://stylelint.io/migration-guide/to-15
-        // indentation: 4,
-        'plugin/no-unused-selectors': {
-            'suffixesToStrip': ['.module'],
-            'documents': [
-                '{cssDir}/{cssName}.tsx',
-            ],
-        },
-        'csstools/value-no-unknown-custom-properties': [
-            true, {
-                importFrom: cssPaths,
-            },
-        ],
-        'selector-pseudo-class-no-unknown': [
-            true,
-            {
-                ignorePseudoClasses: ['global'],
-            },
-        ],
-    },
-};
