+++ packages/go-ui-storybook/package.json
@@ -46,8 +46,8 @@
         "@storybook/test": "^8.1.5",
         "@types/react": "^18.2.43",
         "@types/react-dom": "^18.2.17",
-        "@typescript-eslint/eslint-plugin": "^8.11.0",
-        "@typescript-eslint/parser": "^8.11.0",
+        "@typescript-eslint/eslint-plugin": "^8.24.1",
+        "@typescript-eslint/parser": "^8.24.1",
         "@vitejs/plugin-react-swc": "^3.5.0",
         "chromatic": "^11.4.1",
         "cross-var": "^1.1.0",
