+++ packages/ui/package.json
@@ -71,8 +71,8 @@
         "@types/react": "^18.0.28",
         "@types/react-dom": "^18.0.11",
         "@types/sanitize-html": "^2.9.0",
-        "@typescript-eslint/eslint-plugin": "^8.11.0",
-        "@typescript-eslint/parser": "^8.11.0",
+        "@typescript-eslint/eslint-plugin": "^8.24.1",
+        "@typescript-eslint/parser": "^8.24.1",
         "@vitejs/plugin-react-swc": "^3.5.0",
         "autoprefixer": "^10.4.16",
         "eslint": "^9.20.1",
