+++ packages/ui/src/components/ChartAxes/styles.module.css
@@ -16,7 +16,7 @@
         justify-content: flex-end;
         padding: 0 var(--go-ui-spacing-2xs);
         text-align: right;
-        word-break: break-word;
+        overflow-wrap: break-word;
         font-size: var(--go-ui-font-size-xs);
     }
 
@@ -25,7 +25,7 @@
             transition: var(--go-ui-duration-transition-medium) ease-in-out transform;
             padding: var(--go-ui-spacing-2xs) 0;
             text-align: center;
-            word-break: break-word;
+            overflow-wrap: break-word;
             font-size: var(--go-ui-font-size-xs);
 
             @media screen and (max-width: 30rem) {
