+++ packages/ui/src/components/Table/ColumnShortcuts/styles.module.css
@@ -1,5 +1,5 @@
 .title {
-    color: inherit,
+    color: inherit;
 }
 
 .expansion-indicator-cell-container {
