+++ packages/ui/src/components/Table/TableBodyContent/styles.module.css
@@ -2,7 +2,7 @@
     .cell {
         padding: var(--go-ui-spacing-sm);
         overflow: hidden;
-        word-break: break-word;
+        overflow-wrap: break-word;
     }
 
     &:hover {
