+++ packages/ui/src/components/printable/DescriptionText/styles.module.css
@@ -1,5 +1,5 @@
 .description-text {
     text-align: justify;
     white-space: pre-wrap;
-    word-break: break-word;
+    overflow-wrap: break-word;
 }
